Don't delete DLL for ngspice version <=31
"this would be a good place to add one more fix for when user specifies
ngspice version <=31. In those versions, the DLL naming did not contain
a version number. …  if downloaded DLL's name is already 'ngspice.dll',
the unlink command causes the DLL to be deleted, and results in error.
… the code below adds a check to prevent deletion if the dll-vs
subdirectory only contains 'ngspice.dll'."

https://github.com/FabriceSalvaire/PySpice/pull/270#pullrequestreview-627574464

+++ PySpice/Scripts/pyspice_post_installation.py
@@ -228,17 +228,18 @@ class PySpicePostInstallation:
         # src = dll_path.joinpath('ngspice-{}.dll'.format(self.ngspice_version))
         src = 'ngspice-{}.dll'.format(self.ngspice_version)
         target = dll_path.joinpath('ngspice.dll')
-        if target.exists():
-            target.unlink()
-        try:
-            target.symlink_to(src)
-        except OSError:
-            # OSError: symbolic link privilege not held
-            # Windows: If User Account Control (UAC) is on, any user with the "Create Symbolic
-            #   Links" privilege that is not in the Administrators group can simply create a
-            #   symbolic link.  For users within the Administrators group and with UAC on, the user
-            #   must "Run as Administrator".
-            shutil.copy(target.parent.joinpath(src), target)
+        if dll_path.joinpath(src).exists():
+            if target.exists():
+                target.unlink()
+            try:
+                target.symlink_to(src)
+            except OSError:
+                # OSError: symbolic link privilege not held
+                # Windows: If User Account Control (UAC) is on, any user with the "Create Symbolic
+                #   Links" privilege that is not in the Administrators group can simply create a
+                #   symbolic link.  For users within the Administrators group and with UAC on, the user
+                #   must "Run as Administrator".
+                shutil.copy(target.parent.joinpath(src), target)
 
         spinit_path = spice64_path.joinpath('share', 'ngspice', 'scripts', 'spinit')
         with open(spinit_path) as fh:
